@@ -14,18 +14,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~8.2.14",
+    "@angular/animations": "~9.1.13",
     "@angular/cdk": "~8.2.3",
-    "@angular/common": "~8.2.14",
-    "@angular/compiler": "~8.2.14",
-    "@angular/core": "~8.2.14",
+    "@angular/common": "~9.1.13",
+    "@angular/compiler": "~9.1.13",
+    "@angular/core": "~9.1.13",
     "@angular/flex-layout": "~8.0.0-beta.27",
-    "@angular/forms": "~8.2.14",
+    "@angular/forms": "~9.1.13",
     "@angular/material": "~8.2.3",
     "@angular/material-moment-adapter": "~8.2.3",
-    "@angular/platform-browser": "~8.2.14",
-    "@angular/platform-browser-dynamic": "~8.2.14",
-    "@angular/router": "~8.2.14",
+    "@angular/platform-browser": "~9.1.13",
+    "@angular/platform-browser-dynamic": "~9.1.13",
+    "@angular/router": "~9.1.13",
     "@ngx-translate/core": "~12.1.2",
     "@ngx-translate/http-loader": "~4.0.0",
     "angular-resizable-element": "3.3.4",
@@ -33,22 +33,22 @@
     "moment": "2.29.2",
     "ngx-material-timepicker": "5.4.3",
     "reflect-metadata": "0.1.13",
-    "rxjs": "~6.4.0",
-    "tslib": "~1.10.0",
-    "zone.js": "0.9.1"
+    "rxjs": "~6.6.7",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.22",
-    "@angular-devkit/build-ng-packagr": "~0.803.22",
+    "@angular-devkit/build-angular": "~0.901.15",
+    "@angular-devkit/build-ng-packagr": "~0.901.15",
     "@angular-devkit/build-optimizer": "~0.803.23",
-    "@angular/cli": "~8.3.21",
-    "@angular/compiler-cli": "~8.2.14",
-    "@angular/language-service": "~8.2.14",
+    "@angular/cli": "~9.1.15",
+    "@angular/compiler-cli": "~9.1.13",
+    "@angular/language-service": "~9.1.13",
     "@types/hammerjs": "~2.0.36",
     "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "5.2.1",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "copyfiles": "~2.2.0",
     "jasmine-core": "3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
@@ -58,12 +58,11 @@
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "~1.4.0",
     "karma-sonarqube-unit-reporter": "0.0.23",
-    "ng-packagr": "4.1.0",
+    "ng-packagr": "^9.0.0",
     "protractor": "~5.4.0",
     "scss-bundle": "3.1.1",
     "sonarqube-scanner": "^2.8.1",
     "ts-node": "~7.0.0",
-    "tsickle": "~0.37.0",
     "tslint": "~5.15.0",
     "typescript": "~3.8.3",
     "google-libphonenumber": "^3.2.3"
