@@ -121,22 +121,22 @@ export class OTableMenuComponent implements OTableMenu, OnInit, AfterViewInit, O
 
   @ViewChild('menu', { static: true })
   matMenu: MatMenu;
-  @ViewChild('selectAllCheckboxOption', { static: false })
+  @ViewChild('selectAllCheckboxOption')
   selectAllCheckboxOption: OTableOptionComponent;
-  @ViewChild('exportButtonOption', { static: false })
+  @ViewChild('exportButtonOption')
   exportButtonOption: OTableOptionComponent;
-  @ViewChild('columnsVisibilityButtonOption', { static: false })
+  @ViewChild('columnsVisibilityButtonOption')
   columnsVisibilityButtonOption: OTableOptionComponent;
-  @ViewChild('filterMenuButton', { read: ElementRef, static: false })
+  @ViewChild('filterMenuButton', { read: ElementRef })
   filterMenuButton: ElementRef;
-  @ViewChild('configurationMenuButton', { read: ElementRef, static: false })
+  @ViewChild('configurationMenuButton', { read: ElementRef })
   configurationMenuButton: ElementRef;
 
-  @ViewChild('filterMenu', { static: false })
+  @ViewChild('filterMenu')
   filterMenu: MatMenu;
-  @ViewChild('configurationMenu', { static: false })
+  @ViewChild('configurationMenu')
   configurationMenu: MatMenu;
-  @ViewChild('columnFilterOption', { static: false })
+  @ViewChild('columnFilterOption')
   columnFilterOption: OTableOptionComponent;
   @ViewChild('chartMenu', { static: true })
   chartMenu: MatMenu;
